// apps/zombie-map/src/app/events/area.enum.ts
export enum Area {
    Background = 'background',
    Water = 'water',
    LandcoverIceShelf = 'landcover_ice_shelf',
    LandcoverGlacier = 'landcover_glacier',
    LanduseResidential = 'landuse_residential',
    LandcoverWood = 'landcover_wood',
    LandusePark = 'landuse_park',
    Waterway = 'waterway',
    WaterName = 'water_name',
    Building = 'building',
    AerowayTaxiway = 'aeroway-taxiway',
    AerowayRunwayCasing = 'aeroway-runway-casing',
    AerowayArea = 'aeroway-area',
    AerowayRunway = 'aeroway-runway',
    HighwayPath = 'highway_path',
    HighwayMinor = 'highway_minor',
    HighwayMajorCasing = 'highway_major_casing',
    HighwayMajorInner = 'highway_major_inner',
    HighwayMajorSubtle = 'highway_major_subtle',
    HighwayMotorwayCasing = 'highway_motorway_casing',
    HighwayMotorwayInner = 'highway_motorway_inner',
    HighwayMotorwaySubtle = 'highway_motorway_subtle',
    RailwayTransit = 'railway_transit',
    RailwayTransitDashline = 'railway_transit_dashline',
    RailwayMinor = 'railway_minor',
    RailwayMinorDashline = 'railway_minor_dashline',
    Railway = 'railway',
    RailwayDashline = 'railway_dashline',
    HighwayNameOther = 'highway_name_other',
    HighwayNameMotorway = 'highway_name_motorway',
    BoundaryState = 'boundary_state',
    BoundaryCountry = 'boundary_country',
    PlaceOther = 'place_other',
    PlaceSuburb = 'place_suburb',
    PlaceVillage = 'place_village',
    PlaceTown = 'place_town',
    PlaceCity = 'place_city',
    PlaceCityLarge = 'place_city_large',
    PlaceState = 'place_state',
    PlaceCountryOther = 'place_country_other',
    PlaceCountryMinor = 'place_country_minor',
    PlaceCountryMajor = 'place_country_major'
  }
  